+++ src/app/features/events/events-routing.module.ts
@@ -10,6 +10,7 @@ import { EventsComponent } from '@features/events/events.component';
 import { EventsFavComponent } from '@features/events/events-fav/events-fav.component';
 import { EventsAuditComponent } from '@features/events/events-audit/events-audit.component';
 import { EventsAuditSocialComponent } from '@features/events/events-audit-social/events-audit-social.component';
+import { EventsEntitiesComponent } from '@features/events/events-entities/events-entities.component';
 import { EventsOwnComponent } from '@features/events/events-own/events-own.component';
 import { EventViewComponent } from '@features/events/event-view/event-view.component';
 import { EventConfigComponent } from '@features/events/event-config/event-config.component';
@@ -52,6 +53,12 @@ const routes: Routes = [
   {
     path: 'propios',
     component: EventsOwnComponent,
+    canActivate: [ AngularFireAuthGuard ],
+    data: { authGuardPipe: redirectUnauthorizedToLogin },
+  },
+  {
+    path: 'entidades',
+    component: EventsEntitiesComponent,
     canActivate: [ AngularFireAuthGuard ],
     data: { authGuardPipe: redirectUnauthorizedToLogin },
   },