+++ src/app/shared/shared.module.ts
@@ -7,7 +7,7 @@ import { MaterialModule } from '@shared/material/material.module';
 
 import { SearchFilterPipe } from '@pipes/search-filter.pipe';
 import { StatusPipe } from '@pipes/status.pipe';
-import { EventTypePipe } from '@pipes/eventType.pipe';
+import { EventModePipe } from '@pipes/eventMode.pipe';
 
 import { RoleDirective } from '@shared/directives/role.directive';
 import { ShellComponent } from '@shared/layout/shell/shell.component';
@@ -50,7 +50,7 @@ const components = [
 const pipes = [
   SearchFilterPipe,
   StatusPipe,
-  EventTypePipe,
+  EventModePipe,
 ];
 
 const directives = [